requirements.txt
acme==2.6.0
aiofiles==23.2.1
aiohttp==3.8.5
aiomysql==0.2.0
aioredis==2.0.1
aioschedule==0.5.2
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certbot==2.6.0
certbot-nginx==2.6.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
ConfigArgParse==1.7
configobj==5.0.8
cryptography==41.0.3
distro==1.8.0
exceptiongroup==1.1.2
frozenlist==1.4.0
h11==0.14.0
python-multipart==0.0.6
idna==3.4
josepy==1.13.0
multidict==6.0.4
openai==1.5.0
parsedatetime==2.6
psutil==5.9.5
pycparser==2.21
pydantic==1.10.9
pydantic-core==2.4.0
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyRFC3339==1.1
pyTelegramBotAPI==4.14.1
python-stdnum==1.18
python-telegram-bot==20.4
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.31.0
supervisor==4.2.5
tqdm==4.66.1
typing-extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
watchdog==3.0.0
yarl==1.9.2
Pillow==10.1.0
